+++ admin/settings.php
@@ -21,7 +21,18 @@ $data 		= getXML(GSUSERSPATH . $file);
 $USR 			= stripslashes($data->USR);
 $PASSWD 	= $data->PWD;
 $EMAIL 		= $data->EMAIL;
-$pwd1 =null;$error =null;$success=null;$pwd2 =null;$editorchck =null; $prettychck =null;
+$NAME			= $data->NAME;
+
+$lang_array = getFiles(GSLANGPATH);
+
+# initialize these all as null
+$pwd1 = $error = $success = $pwd2 = $editorchck = $prettychck = null;
+
+# if the flush cache command was invoked
+if (isset($_GET['flushcache'])) { 
+	delete_cache();
+	$update = 'flushcache-success';
+}
 
 # if the undo command was invoked
 if (isset($_GET['undo'])) { 
@@ -68,8 +79,8 @@ if(isset($_POST['submitted'])) {
 		$SITEURL = tsl($_POST['siteurl']); 
 	}
 	if(isset($_POST['permalink'])) { 
-		$PERMALINK = $_POST['permalink']; 
-	}	
+		$PERMALINK = trim($_POST['permalink']); 
+	}
 	if(isset($_POST['template'])) { 
 		$TEMPLATE = $_POST['template']; 
 	}
@@ -82,18 +93,21 @@ if(isset($_POST['submitted'])) {
 	# user-specific fields
 	if(isset($_POST['user'])) { 
 		$USR = strtolower($_POST['user']); 
+	}
+ 	if(isset($_POST['name'])) { 
+		$NAME = $_POST['name']; 
 	} 
 	if(isset($_POST['email'])) { 
 		$EMAIL = $_POST['email']; 
 	} 
 	if(isset($_POST['timezone'])) { 
-		$TIMEZONE = $_POST['timezone']; 
+		$TIMEZONE = var_out($_POST['timezone']); 
 	}
 	if(isset($_POST['lang'])) { 
-		$LANG = $_POST['lang']; 
+		$LANG = var_out($_POST['lang']); 
 	}
 	if(isset($_POST['show_htmleditor'])) {
-	  $HTMLEDITOR = $_POST['show_htmleditor']; 
+	  $HTMLEDITOR = var_out($_POST['show_htmleditor']); 
 	} else {
 		$HTMLEDITOR = '';
 	}
@@ -111,13 +125,17 @@ if(isset($_POST['submitted'])) {
 			$PASSWD = passhash($pwd1); 
 		}	
 		
+		// check valid lang files
+		if(!in_array($LANG.'.php', $lang_array) and !in_array($LANG.'.PHP', $lang_array)) die(); 
+
 		# create user xml file
 		createBak($file, GSUSERSPATH, GSBACKUSERSPATH);
 		if (file_exists(GSUSERSPATH . _id($USR).'.xml.reset')) { unlink(GSUSERSPATH . _id($USR).'.xml.reset'); }	
 		$xml = new SimpleXMLElement('<item></item>');
 		$xml->addChild('USR', $USR);
+		$xml->addChild('NAME', var_out($NAME));
 		$xml->addChild('PWD', $PASSWD);
-		$xml->addChild('EMAIL', $EMAIL);
+		$xml->addChild('EMAIL', var_out($EMAIL,'email'));
 		$xml->addChild('HTMLEDITOR', $HTMLEDITOR);
 		$xml->addChild('TIMEZONE', $TIMEZONE);
 		$xml->addChild('LANG', $LANG);
@@ -138,7 +156,7 @@ if(isset($_POST['submitted'])) {
 		$note = $xmls->addChild('TEMPLATE');
 		$note->addCData($TEMPLATE);
 		$xmls->addChild('PRETTYURLS', $PRETTYURLS);
-		$xmls->addChild('PERMALINK', $PERMALINK);
+		$xmls->addChild('PERMALINK', var_out($PERMALINK));
 		
 		exec_action('settings-website');
 		
@@ -162,21 +180,16 @@ if ($HTMLEDITOR != '' ) { $editorchck = 'checked'; }
 if ($PRETTYURLS != '' ) { $prettychck = 'checked'; }
 
 # get all available language files
-$lang_handle = opendir(GSLANGPATH) or die("Unable to open ". GSLANGPATH);
 if ($LANG == ''){ $LANG = 'en_US'; }
-while ($lfile = readdir($lang_handle)) {
-	if( is_file(GSLANGPATH . $lfile) && $lfile != "." && $lfile != ".." )	{
-		$lang_array[] = basename($lfile, ".php");
-	}
-}
+
 if (count($lang_array) != 0) {
 	sort($lang_array);
-	$count = '0'; $sel = ''; $langs = '';
-	foreach ($lang_array as $larray){
-		if ($LANG == $larray)	{ $sel="selected"; }
-		$langs .= '<option '.$sel.' value="'.$larray.'" >'.$larray.'</option>';
+	$sel = ''; $langs = '';
+	foreach ($lang_array as $lfile){
+		$lfile = basename($lfile,".php");
+		if ($LANG == $lfile)	{ $sel="selected"; }
+		$langs .= '<option '.$sel.' value="'.$lfile.'" >'.$lfile.'</option>';
 		$sel = '';
-		$count++;
 	}
 } else {
 	$langs = '<option value="" selected="selected" >-- '.i18n_r('NONE').' --</option>';
@@ -209,7 +222,8 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('GENERAL_SETTINGS'));
 		<p class="inline" ><input name="prettyurls" id="prettyurls" type="checkbox" value="1" <?php echo $prettychck; ?>  /> &nbsp;<label for="prettyurls" ><?php i18n('USE_FANCY_URLS');?></label></p>
 				
 		<div class="leftsec">
-			<p><label for="permalink"  class="clearfix"><?php i18n('PERMALINK');?>: <span class="right"><a href="http://get-simple.info/wiki/pretty_urls" target="_blank" ><?php i18n('MORE');?></a></span></label><input class="text" name="permalink" id="permalink" type="text" value="<?php if(isset($PERMALINK)) { echo $PERMALINK; } ?>" /></p>
+			<p><label for="permalink"  class="clearfix"><?php i18n('PERMALINK');?>: <span class="right"><a href="http://get-simple.info/docs/pretty_urls" target="_blank" ><?php i18n('MORE');?></a></span></label><input class="text" name="permalink" id="permalink" type="text" placeholder="%parent%/%slug%/" value="<?php if(isset($PERMALINK)) { echo var_out($PERMALINK); } ?>" /></p>
+		<a id="flushcache" class="button" href="?flushcache"><?php i18n('FLUSHCACHE'); ?></a>
 		</div>
 		<div class="clear"></div>
 		
@@ -223,12 +237,18 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('GENERAL_SETTINGS'));
 			<p><label for="user" ><?php i18n('LABEL_USERNAME');?>:</label><input class="text" id="user" name="user" type="text" readonly value="<?php if(isset($USR1)) { echo $USR1; } else { echo $USR; } ?>" /></p>
 		</div>
 		<div class="rightsec">
-			<p><label for="email" ><?php i18n('LABEL_EMAIL');?>:</label><input class="text" id="email" name="email" type="email" value="<?php if(isset($EMAIL1)) { echo $EMAIL1; } else { echo $EMAIL; } ?>" /></p>
+			<p><label for="email" ><?php i18n('LABEL_EMAIL');?>:</label><input class="text" id="email" name="email" type="email" value="<?php if(isset($EMAIL1)) { echo $EMAIL1; } else { echo var_out($EMAIL,'email'); } ?>" /></p>
 			<?php if (! check_email_address($EMAIL)) {
 				echo '<p style="margin:-15px 0 20px 0;color:#D94136;font-size:11px;" >'.i18n_r('WARN_EMAILINVALID').'</p>';
 			}?>
 		</div>
 		<div class="clear"></div>
+		<div class="leftsec">
+			<p><label for="name" ><?php i18n('LABEL_DISPNAME');?>:</label>
+			<span style="margin:0px 0 5px 0;font-size:12px;color:#999;" ><?php i18n('DISPLAY_NAME');?></span>			
+			<input class="text" id="name" name="name" type="text" value="<?php if(isset($NAME1)) { echo $NAME1; } else { echo var_out($NAME); } ?>" /></p>
+		</div>		
+		<div class="clear"></div>		
 		<div class="leftsec">
 			<p><label for="timezone" ><?php i18n('LOCAL_TIMEZONE');?>:</label>
 			<?php if( (isset($_POST['timezone'])) ) { $TIMEZONE = $_POST['timezone']; } ?>
@@ -239,7 +259,7 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('GENERAL_SETTINGS'));
 			</p>
 		</div>
 		<div class="rightsec">
-			<p><label for="lang" ><?php i18n('LANGUAGE');?>: <span class="right"><a href="http://get-simple.info/wiki/languages" target="_blank" ><?php i18n('MORE');?></a></span></label>
+			<p><label for="lang" ><?php i18n('LANGUAGE');?>: <span class="right"><a href="http://get-simple.info/docs/languages" target="_blank" ><?php i18n('MORE');?></a></span></label>
 			<select name="lang" id="lang" class="text">
 				<?php echo $langs; ?>
 			</select>
